oh-tab-abc: dedupe welcome secret-status
