+++ src/extension.ts
@@ -29,6 +29,7 @@ import { createHalConfigurationChangeHandler } from './extension/halConfiguratio
 import { formatEnvironmentInformation } from './shared/environmentInformation';
 import { getFileBackedFsPath } from './shared/uri';
 import { resolvePreferredWorkspaceRoot } from './shared/workspaceRoot';
+import { computeWelcomeSecretStatus } from './shared/welcomeSecretStatus';
 import { getServerSessionApiKeySecretKey } from './auth/serverSessionApiKeys';
 import { registerCloudLoginCommand } from './extension/cloudLoginCommand';
 import { registerCloudLogoutCommand } from './extension/cloudLogoutCommand';
@@ -347,65 +348,11 @@ export function activate(context: vscode.ExtensionContext) {
   let lastWelcomeSecretStatus: { hasProviderKey: boolean; hasGeminiKey: boolean } | null = null;
   let welcomeSecretStatusTimer: ReturnType<typeof setTimeout> | null = null;
 
-  const computeWelcomeSecretStatus = async (): Promise<{ hasProviderKey: boolean; hasGeminiKey: boolean }> => {
-    const settingsMgr = new SettingsManager(new VscodeSettingsAdapter(context));
-    const settings = await settingsMgr.get();
-
-    const secretIsSet = async (key: string): Promise<boolean> => {
-      try {
-        const value = await context.secrets.get(key);
-        return typeof value === 'string' && value.trim().length > 0;
-      } catch {
-        return false;
-      }
-    };
-
-    const envIsSet = (key: string): boolean => {
-      const value = process.env[key];
-      return typeof value === 'string' && value.trim().length > 0;
-    };
-
-    const hasGeminiKey =
-      (await secretIsSet('GEMINI_API_KEY')) ||
-      envIsSet('GEMINI_API_KEY') ||
-      (settings.llm.provider === 'gemini' && typeof settings.secrets.llmApiKey === 'string' && settings.secrets.llmApiKey.trim().length > 0);
-
-    const hasGenericKey = typeof settings.secrets.llmApiKey === 'string' && settings.secrets.llmApiKey.trim().length > 0;
-
-    const providerStorageKeys = [
-      'OPENAI_API_KEY',
-      'ANTHROPIC_API_KEY',
-      'OPENROUTER_API_KEY',
-      'LITELLM_API_KEY',
-    ];
-
-    let hasAnyProviderStorageKey = false;
-    for (const key of providerStorageKeys) {
-      if (envIsSet(key) || await secretIsSet(key)) {
-        hasAnyProviderStorageKey = true;
-        break;
-      }
-    }
-
-    let hasAnyProfileKey = false;
-    try {
-      for (const profileId of listProfiles()) {
-        if (await secretIsSet(`openhands.llmProfileApiKey.${profileId}`)) {
-          hasAnyProfileKey = true;
-          break;
-        }
-      }
-    } catch {
-      hasAnyProfileKey = false;
-    }
-
-    const hasProviderKey = hasGeminiKey || hasGenericKey || hasAnyProviderStorageKey || hasAnyProfileKey;
-    return { hasProviderKey, hasGeminiKey };
-  };
-
   const postWelcomeSecretStatus = async (): Promise<void> => {
     if (!chatView) return;
-    const status = await computeWelcomeSecretStatus();
+    const settingsMgr = new SettingsManager(new VscodeSettingsAdapter(context));
+    const settings = await settingsMgr.get();
+    const status = await computeWelcomeSecretStatus({ context, settings });
     if (
       lastWelcomeSecretStatus &&
       lastWelcomeSecretStatus.hasProviderKey === status.hasProviderKey &&
