+++ src/webview/host/handlers/llmProfiles.ts
@@ -1,7 +1,7 @@
 import type * as vscode from 'vscode';
 import type { ConversationInstance } from '@openhands/agent-sdk-ts';
 import { assertValidProfileId, detectProviderFromBaseUrl, LLMProfileValidationError, type LLMProvider } from '@openhands/agent-sdk-ts';
-import type { OpenHandsSettings, SettingsManager } from '../../../settings/SettingsManager';
+import type { SettingsManager } from '../../../settings/SettingsManager';
 import { resolveConfiguredLlmLabel } from '../../../shared/llmProfiles';
 import { STATUS_MESSAGE_DISMISS_DELAY_MS, type WebviewToHostMessage } from '../../../shared/webviewMessages';
 import type { CreateWebviewMessageHandlerDeps, WebviewHost } from '../createWebviewMessageHandler';
@@ -366,59 +366,4 @@ export async function handleLlmProfileApiKeySetRequest(args: {
     const reason = err instanceof Error ? err.message : String(err);
     void args.host.postMessage({ type: 'llmProfileApiKeySetResponse', requestId, ok: false, profileId, error: reason });
   }
-}
-
-export async function computeWelcomeSecretStatus(args: {
-  deps: CreateWebviewMessageHandlerDeps;
-  context: vscode.ExtensionContext;
-  settings: OpenHandsSettings;
-}): Promise<{ hasProviderKey: boolean; hasGeminiKey: boolean }> {
-  const secretIsSet = async (key: string): Promise<boolean> => {
-    try {
-      const value = await args.context.secrets.get(key);
-      return typeof value === 'string' && value.trim().length > 0;
-    } catch {
-      return false;
-    }
-  };
-
-  const envIsSet = (key: string): boolean => {
-    const value = process.env[key];
-    return typeof value === 'string' && value.trim().length > 0;
-  };
-
-  const hasGeminiKey = (await secretIsSet('GEMINI_API_KEY'))
-    || envIsSet('GEMINI_API_KEY')
-    || (args.settings.llm.provider === 'gemini' && typeof args.settings.secrets.llmApiKey === 'string' && args.settings.secrets.llmApiKey.trim().length > 0);
-
-  const hasGenericKey = typeof args.settings.secrets.llmApiKey === 'string' && args.settings.secrets.llmApiKey.trim().length > 0;
-
-  const providerStorageKeys = [
-    'OPENAI_API_KEY',
-    'ANTHROPIC_API_KEY',
-    'OPENROUTER_API_KEY',
-    'LITELLM_API_KEY',
-  ];
-  let hasAnyProviderStorageKey = false;
-  for (const key of providerStorageKeys) {
-    if (envIsSet(key) || await secretIsSet(key)) {
-      hasAnyProviderStorageKey = true;
-      break;
-    }
-  }
-
-  let hasAnyProfileKey = false;
-  try {
-    for (const profileId of listAvailableLlmProfiles({ deps: args.deps, outputChannel: undefined })) {
-      if (await secretIsSet(`openhands.llmProfileApiKey.${profileId}`)) {
-        hasAnyProfileKey = true;
-        break;
-      }
-    }
-  } catch {
-    hasAnyProfileKey = false;
-  }
-
-  const hasProviderKey = hasGeminiKey || hasGenericKey || hasAnyProviderStorageKey || hasAnyProfileKey;
-  return { hasProviderKey, hasGeminiKey };
 }
